+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Added
   * Added example to `terra-modal-manager` to show how to consume the new closeOnOutsideClick property.
+  
+* Changed
+  * Updated the `terra-date-time-picker` example for field label.
 
 ## 1.64.0 - (February 7, 2024)